update the import path of my ProfilePage
